Reduce test file line limit from 800 to 750
- Updated ESLint config to reduce max-lines for test files from 800 to 750
- Split 4 test files that violated the new limit:
  - barbeat-interpreter-copy.test.js → split range tests to barbeat-interpreter-copy-range.test.js
  - barbeat-interpreter-core.test.js → split time signature tests to barbeat-interpreter-timesig.test.js
  - read-track-devices.test.js → split rack edge cases to read-track-devices-racks.test.js
  - use-settings.test.ts → split provider-specific tests to use-settings-providers.test.ts
- All original files now under 750 lines
- All new test files maintain same test coverage
- All checks pass (lint, typecheck, format, tests)
