+++ config/eslint.config.js
@@ -430,7 +430,7 @@ export default [
       "max-lines": [
         "error",
         {
-          max: 800,
+          max: 750,
         },
       ],
     },
